+++ test/support/sample_live.ex
@@ -5,18 +5,12 @@ defmodule SampleLive.Some.NestedPlug do
   def call(socket) do
     socket
   end
-
-  @impl true
-  def call(socket, {_action, _params}) do
-    socket
-  end
 end
 
 defmodule Reusable do
   defmacro __using__(_) do
     quote do
       plug SampleLive.Some.NestedPlug when action == :nonexisting
-      plug SampleLive.Some.NestedPlug, {action, params}
     end
   end
 end
@@ -27,9 +21,6 @@ defmodule SampleLive do
   use Reusable
 
   defmodule BeforeGlobal do
-    @behaviour Phoenix.LiveController.Plug
-
-    @impl true
     def call(socket, {name, payload}) do
       assign(socket, :global_plug_called, {name, payload})
     end
@@ -46,14 +37,41 @@ defmodule SampleLive do
       else: assign(socket, user: session["user"])
   end
 
-  plug BeforeGlobal, {action || event || message, params || payload}
+  plug BeforeGlobal.call(socket, {name, params || message})
   plug NestedPlug
 
   @skip_action :index_with_opts
-  plug {BeforeGlobal, :other}, :arg when action != @skip_action and !message
+  plug BeforeGlobal.other(socket, :arg) when action != @skip_action and !message
 
-  plug :before_action_handler, %{p: params, key: :before_action_handler_called} when action
-  plug :before_action_handler, %{p: params, key: :before_action_handler_called_two} when action
+  plug on_final_mount(socket)
+       when action && connected?(socket) && !mounted?(socket) && local_check?(socket)
+
+  defp on_final_mount(socket) do
+    if Map.get(socket.assigns, :final_mount_done) do
+      raise "mounting twice?"
+    else
+      assign(socket, :final_mount_done, true)
+    end
+  end
+
+  defp local_check?(_socket) do
+    true
+  end
+
+  plug :on_final_mount_atom when action && connected?(socket) && !mounted?(socket)
+
+  defp on_final_mount_atom(socket) do
+    if Map.get(socket.assigns, :final_mount_done_2) do
+      raise "mounting twice?"
+    else
+      assign(socket, :final_mount_done_2, true)
+    end
+  end
+
+  plug before_action_handler(socket, %{p: params, key: :before_action_handler_called}) when action
+
+  plug before_action_handler(socket, %{p: params, key: :before_action_handler_called_two})
+       when action
 
   defp before_action_handler(socket, %{p: params, key: key}) do
     history = Map.get(socket.assigns, :plug_history, [])
@@ -63,7 +81,7 @@ defmodule SampleLive do
       else: assign(socket, key, true) |> assign(:plug_history, history ++ [key])
   end
 
-  plug :before_event_handler, params when event
+  plug before_event_handler(socket, params) when event
 
   def before_event_handler(socket, params) do
     if params["redirect"],
@@ -71,10 +89,10 @@ defmodule SampleLive do
       else: assign(socket, before_event_handler_called: true)
   end
 
-  plug :before_message_handler, payload when message
+  plug before_message_handler(socket, message) when message
 
-  def before_message_handler(socket, payload) do
-    if payload == {:x, :redirect},
+  def before_message_handler(socket, message) do
+    if message == {:x, :redirect},
       do: push_redirect(socket, to: "/"),
       else: assign(socket, before_message_handler_called: true)
   end